moved MaxMindFunction.java to presto-hive
