+++ presto-hive/src/main/java/io/prestosql/plugin/hive/HivePlugin.java
@@ -14,8 +14,11 @@
 package io.prestosql.plugin.hive;
 
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableSet;
 import io.prestosql.spi.Plugin;
 import io.prestosql.spi.connector.ConnectorFactory;
+
+import java.util.Set;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Strings.isNullOrEmpty;
@@ -37,4 +40,10 @@ public class HivePlugin
         return ImmutableList.of(new HiveConnectorFactory(name));
     }
 
+    @Override
+    public Set<Class<?>> getFunctions() {
+        return ImmutableSet.<Class<?>>builder()
+                .add(MaxMindFunction.class)
+                .build();
+    }
 }
